// src/actions/upsert-patient/schema.ts
import { z } from "zod";

export const upsertPatientSchema = z.object({
  id: z.string().uuid().optional(),
  clinicId: z.string().uuid().optional(),
  name: z.string().trim().min(1, { message: "Nome é obrigatório." }),
  email: z.string().email({ message: "E-mail inválido." }),
  phoneNumber: z.string().min(10, { message: "Telefone é obrigatório." }),
  sex: z.enum(["male", "female"], { required_error: "Sexo é obrigatório." }),
});

export type UpsertPatientSchema = z.infer<typeof upsertPatientSchema>;